+++ src/js/widgets/index.js
@@ -15,6 +15,7 @@ import { RecentlyAddedWidget } from './widgets/recently-added.js';
 import { TopRatedWidget } from './widgets/top-rated.js';
 import { RecentlyFinishedWidget } from './widgets/recently-finished.js';
 import { SeriesProgressWidget } from './widgets/series-progress.js';
+import { WishlistWidget } from './widgets/wishlist.js';
 
 // Register all widgets in default order
 widgetRegistry.register(WelcomeWidget);
@@ -23,7 +24,8 @@ widgetRegistry.register(RecentlyAddedWidget);
 widgetRegistry.register(TopRatedWidget);
 widgetRegistry.register(RecentlyFinishedWidget);
 widgetRegistry.register(SeriesProgressWidget);
+widgetRegistry.register(WishlistWidget);
 
 // Export for external use
 export { widgetRegistry, BaseWidget };
-export { WelcomeWidget, CurrentlyReadingWidget, RecentlyAddedWidget, TopRatedWidget, RecentlyFinishedWidget, SeriesProgressWidget };
+export { WelcomeWidget, CurrentlyReadingWidget, RecentlyAddedWidget, TopRatedWidget, RecentlyFinishedWidget, SeriesProgressWidget, WishlistWidget };